# docs/openapi.yaml
openapi: "3.0.3"

info:
  title: JM Voice AI Platform API
  description: |
    One Stop Total Solution — Voice AI + POS + CCTV.
    This is the Living API Contract for Phase 1 Foundation + Phase 2 Relay Bridge.
    All endpoints require a Supabase JWT (Bearer token) to resolve tenant_id.
    Phase 2 frontend agents MUST use this document as their sole source of truth.
    (이 문서는 Phase 2 프론트엔드 에이전트의 유일한 작업 지시서입니다.)
  version: "0.3.0"
  contact:
    email: mchang@jmtechone.com

servers:
  - url: http://localhost:8000
    description: Local development (Vite proxy → FastAPI)
  - url: https://api.jmtechone.com
    description: Production

security:
  - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase JWT. The `sub` claim is used as `tenant_id` for all RLS queries.

  schemas:
    # --- Health ---
    HealthResponse:
      type: object
      required: [status, service]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: jm-voice-ai-platform

    # --- Catalog Navigator (Layer 2) ---
    MenuItemOut:
      type: object
      required: [variant_id, item_id, name, category, price, stock_quantity]
      properties:
        variant_id:
          type: string
          description: Loyverse variant UUID (primary key)
          example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        item_id:
          type: string
          description: Loyverse item UUID
          example: "a1b2c3d4-0000-0000-0000-000000000001"
        name:
          type: string
          example: "Classic Burger"
        category:
          type: string
          example: "Mains"
        price:
          type: integer
          description: Price in US cents
          example: 1200
        stock_quantity:
          type: integer
          example: 50

    # --- Slot Filler (Layer 2) ---
    Intent:
      type: string
      enum: [reservation, order]
      description: Dialog intent type (다이얼로그 의도 타입)

    SlotCheckRequest:
      type: object
      required: [intent, collected]
      properties:
        intent:
          $ref: '#/components/schemas/Intent'
        collected:
          type: object
          description: Key-value pairs of already-collected slot values
          additionalProperties: true
          example:
            customer_name: "John Kim"
            customer_phone: "503-555-1234"

    SlotCheckResult:
      type: object
      required: [missing, next_prompt, complete]
      properties:
        missing:
          type: array
          items:
            type: string
          description: List of slot names still required
          example: ["customer_email", "reservation_date"]
        next_prompt:
          type: string
          nullable: true
          description: Voice prompt to ask for the next missing slot
          example: "And your email address?"
        complete:
          type: boolean
          description: True when all required slots are filled
          example: false

    # --- Relay Bridge (Layer 4) ---
    SolinkEventPayload:
      type: object
      required: [event_type, camera_id, location, timestamp]
      properties:
        event_type:
          type: string
          enum: [motion_detected, door_open, door_close, alarm_triggered]
          description: Type of Solink security event (보안 이벤트 유형)
          example: motion_detected
        camera_id:
          type: string
          description: Unique camera identifier (카메라 고유 식별자)
          example: "cam-001"
        location:
          type: string
          description: Physical location label (물리적 위치 레이블)
          example: "Front Entrance"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 event timestamp (ISO 8601 이벤트 타임스탬프)
          example: "2026-04-25T14:30:00Z"
        metadata:
          type: object
          description: Optional additional event metadata (선택적 추가 이벤트 메타데이터)
          additionalProperties: true
          nullable: true

    RelayAck:
      type: object
      required: [accepted, relay_id, queued_at]
      properties:
        accepted:
          type: boolean
          description: Whether the relay was accepted (릴레이 수락 여부)
          example: true
        relay_id:
          type: string
          format: uuid
          description: Server-generated UUID for this relay job (서버 생성 릴레이 UUID)
          example: "550e8400-e29b-41d4-a716-446655440000"
        queued_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when job was queued (작업 큐 등록 시각)
          example: "2026-04-25T14:30:00.123Z"

    LoyverseOrderPayload:
      type: object
      required: [items]
      properties:
        items:
          type: array
          description: List of order items (주문 항목 목록)
          items:
            type: object
            required: [variant_id, quantity]
            properties:
              variant_id:
                type: string
                description: Loyverse variant UUID (Loyverse 상품 변형 UUID)
                example: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
              quantity:
                type: integer
                minimum: 1
                description: Quantity ordered — must be at least 1 (최소 1 이상)
                example: 2
        table_number:
          type: string
          description: Optional table number (선택적 테이블 번호)
          nullable: true
          example: "Table 5"
        note:
          type: string
          description: Optional order note, max 500 characters (선택적 주문 메모, 최대 500자)
          nullable: true
          maxLength: 500
          example: "No onions please"

    LoyverseOrderAck:
      type: object
      required: [accepted, relay_id, loyverse_receipt_id, queued_at]
      properties:
        accepted:
          type: boolean
          description: Whether the order relay was accepted (주문 릴레이 수락 여부)
          example: true
        relay_id:
          type: string
          format: uuid
          description: Server-generated UUID for this relay job (서버 생성 릴레이 UUID)
          example: "550e8400-e29b-41d4-a716-446655440001"
        loyverse_receipt_id:
          type: string
          nullable: true
          description: Loyverse receipt ID — null until async relay completes (비동기 릴레이 완료 전까지 null)
          example: null
        queued_at:
          type: string
          format: date-time
          description: ISO 8601 timestamp when job was queued (작업 큐 등록 시각)
          example: "2026-04-25T14:30:00.456Z"

    # --- Solink CCTV Read (Layer 4) ---
    CameraItem:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: string
          description: Unique camera identifier (카메라 고유 식별자)
          example: "cam-001"
        name:
          type: string
          description: Human-readable camera name (카메라 표시 이름)
          example: "Front Entrance"
        status:
          type: string
          description: Camera status (카메라 상태)
          example: "online"

    CameraListResponse:
      type: object
      required: [cameras]
      properties:
        cameras:
          type: array
          description: List of cameras from Solink (Solink 카메라 목록)
          items:
            $ref: '#/components/schemas/CameraItem'

    VideoLinkResponse:
      type: object
      required: [url, camera_id, timestamp]
      properties:
        url:
          type: string
          nullable: true
          description: Video playback URL, null if footage unavailable (비디오 재생 URL, 영상 없으면 null)
          example: "https://solink.example.com/playback/cam-001?t=2026-04-25T14:30:00Z"
        camera_id:
          type: string
          description: Camera identifier (카메라 식별자)
          example: "cam-001"
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 timestamp that was requested (요청한 ISO 8601 타임스탬프)
          example: "2026-04-25T14:30:00Z"

    # --- Errors ---
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
          example: "Invalid or expired token"

paths:
  # ── Health ──────────────────────────────────────────────────────────────
  /health:
    get:
      summary: Health check
      description: Returns service liveness status. No auth required.
      operationId: healthCheck
      security: []
      tags: [System]
      responses:
        "200":
          description: Service is alive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  # ── Catalog Navigator (Layer 2) ─────────────────────────────────────────
  /api/catalog/menu:
    get:
      summary: Get live menu for the authenticated tenant
      description: |
        Returns all active menu items for the calling tenant, ordered by
        category then name. The `tenant_id` is extracted from the JWT — never
        passed as a parameter. This enforces Row-Level Security at the API layer.
        (JWT에서 tenant_id를 추출 — RLS 강제)
      operationId: getCatalogMenu
      tags: [Catalog Navigator]
      responses:
        "200":
          description: Menu items for tenant
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItemOut'
              example:
                - variant_id: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                  item_id: "a1b2c3d4-0000-0000-0000-000000000001"
                  name: "Classic Burger"
                  category: "Mains"
                  price: 1200
                  stock_quantity: 50
        "401":
          description: Missing or invalid JWT
          headers:
            WWW-Authenticate:
              schema:
                type: string
                example: Bearer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Slot Filler (Layer 2) ───────────────────────────────────────────────
  /api/slots/check:
    post:
      summary: Check slot completeness for a dialog intent
      description: |
        Given an intent (reservation | order) and the slots already collected
        during conversation, returns which fields are still missing and the
        next voice prompt to deliver. Used by the Gemini orchestration loop.
        (대화 중 수집된 슬롯을 기반으로 누락 필드와 다음 음성 안내를 반환)
      operationId: checkSlots
      tags: [Slot Filler]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlotCheckRequest'
            examples:
              reservation_partial:
                summary: Partial reservation — name collected
                value:
                  intent: reservation
                  collected:
                    customer_name: "John Kim"
              order_complete:
                summary: Order fully collected
                value:
                  intent: order
                  collected:
                    customer_name: "Jane Lee"
                    customer_phone: "503-555-9999"
                    customer_email: "jane@example.com"
                    items: [{"variant_id": "f47ac10b", "quantity": 2}]
                    user_explicit_confirmation: true
      responses:
        "200":
          description: Slot check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotCheckResult'
              examples:
                incomplete:
                  summary: Missing slots remain
                  value:
                    missing: ["customer_phone", "customer_email", "reservation_date", "reservation_time", "party_size"]
                    next_prompt: "What's the best phone number to reach you?"
                    complete: false
                complete:
                  summary: All slots filled
                  value:
                    missing: []
                    next_prompt: null
                    complete: true
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error — invalid intent or malformed body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ── Relay Bridge (Layer 4) ──────────────────────────────────────────────
  /api/relay/solink/event:
    post:
      summary: Forward Solink security event (Fire-and-Forget)
      description: |
        Accepts a Solink security event and relays it asynchronously. Returns 202
        immediately; the actual forwarding happens in the background. JWT tenant_id
        is used for routing.
        (Solink 보안 이벤트를 비동기적으로 수락하고 릴레이. 즉시 202 반환 — 실제 전달은 백그라운드 처리)
      operationId: relaySolinkEvent
      tags: [Relay Bridge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolinkEventPayload'
            example:
              event_type: motion_detected
              camera_id: "cam-001"
              location: "Front Entrance"
              timestamp: "2026-04-25T14:30:00Z"
              metadata:
                confidence: 0.95
      responses:
        "202":
          description: Relay job accepted — forwarding happens asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RelayAck'
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error — invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/relay/loyverse/order:
    post:
      summary: Submit order to Loyverse POS (Fire-and-Forget)
      description: |
        Accepts an order and relays it to Loyverse POS asynchronously. Returns 202
        immediately. All items must reference valid variant_ids from the tenant's
        Loyverse catalog.
        (주문을 받아 Loyverse POS에 비동기적으로 릴레이. 즉시 202 반환 — 모든 항목은 테넌트 카탈로그의 유효한 variant_id 참조 필요)
      operationId: relayLoyverseOrder
      tags: [Relay Bridge]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoyverseOrderPayload'
            example:
              items:
                - variant_id: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                  quantity: 2
              table_number: "Table 5"
              note: "No onions please"
      responses:
        "202":
          description: Order relay accepted — Loyverse forwarding happens asynchronously
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyverseOrderAck'
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "422":
          description: Validation error — invalid payload or missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/relay/solink/cameras:
    get:
      summary: Get list of Solink cameras
      description: |
        Retrieves all cameras registered in the Solink CCTV system for this tenant.
        Requires JWT auth; tenant_id is extracted from the token for RLS enforcement.
        (Solink CCTV 시스템에서 테넌트 카메라 목록 조회. JWT 인증 필요; tenant_id로 RLS 강제)
      operationId: getSolinkCameras
      tags: [Relay Bridge]
      responses:
        "200":
          description: Camera list retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CameraListResponse'
              example:
                cameras:
                  - id: "cam-001"
                    name: "Front Entrance"
                    status: "online"
                  - id: "cam-002"
                    name: "Back Office"
                    status: "offline"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/relay/solink/video-link:
    get:
      summary: Get video playback link for a camera at a timestamp
      description: |
        Returns a video playback URL from Solink for the given camera and timestamp.
        If no footage exists for that moment, url is null (not an error condition).
        (지정된 카메라와 타임스탬프의 비디오 재생 링크 반환. 영상 없으면 url이 null — 에러 아님)
      operationId: getSolinkVideoLink
      tags: [Relay Bridge]
      parameters:
        - name: camera_id
          in: query
          required: true
          schema:
            type: string
          description: Camera identifier (카메라 식별자)
          example: "cam-001"
        - name: timestamp
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: ISO 8601 timestamp (ISO 8601 타임스탬프)
          example: "2026-04-25T14:30:00Z"
      responses:
        "200":
          description: Video link result — url may be null if no footage available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoLinkResponse'
              examples:
                with_footage:
                  summary: Footage available
                  value:
                    url: "https://solink.example.com/playback/cam-001?t=2026-04-25T14:30:00Z"
                    camera_id: "cam-001"
                    timestamp: "2026-04-25T14:30:00Z"
                no_footage:
                  summary: No footage at that moment
                  value:
                    url: null
                    camera_id: "cam-001"
                    timestamp: "2026-04-25T14:30:00Z"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/relay/solink/snapshot:
    get:
      summary: Get JPEG snapshot for a camera at a timestamp
      description: |
        Returns a JPEG image snapshot from Solink for the given camera and timestamp.
        If no snapshot is available, returns 200 with JSON error body.
        (지정된 카메라와 타임스탬프의 JPEG 스냅샷 반환. 스냅샷 없으면 JSON 에러 본문으로 200 반환)
      operationId: getSolinkSnapshot
      tags: [Relay Bridge]
      parameters:
        - name: camera_id
          in: query
          required: true
          schema:
            type: string
          description: Camera identifier (카메라 식별자)
          example: "cam-001"
        - name: timestamp
          in: query
          required: true
          schema:
            type: string
            format: date-time
          description: ISO 8601 timestamp (ISO 8601 타임스탬프)
          example: "2026-04-25T14:30:00Z"
      responses:
        "200":
          description: JPEG snapshot or unavailability notice
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "snapshot unavailable"
        "401":
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

tags:
  - name: System
    description: Infrastructure and health endpoints
  - name: Catalog Navigator
    description: Layer 2 — Live menu retrieval with RLS-enforced tenant isolation
  - name: Slot Filler
    description: Layer 2 — Dialog slot collection and completeness checking
  - name: Relay Bridge
    description: Layer 4 — Fire-and-Forget external bridges to Solink CCTV and Loyverse POS